Small fixes and updates around various providers (#2717)
* Update DDG.py: added Llama 3.3 Instruct and o3-mini

Duck.ai now supports o3-mini, and previous Llama 3.1 70B is now replaced by Llama 3.3 70B.

* Update DDG.py: change Llama 3.3 70B Instruct ID to "meta-llama/Llama-3.3-70B-Instruct-Turbo"

Fixed typo in full model name

* Update Cerebras.py: add "deepseek-r1-distill-llama-70b" to the models list

Cerebras now provides inference for a DeepSeek-R1 distilled to Llama 3.3 70B as well.

* Update models.py: reflect changes in DDG provider

- Removed DDG from best providers list for Llama 3.1 70B
- Added DDG to best providers list for o3-mini and Llama 3.3 70B

* A small update in HuggingFaceInference get_models() method

Previously, get_models() method was returning "TypeError: string indices must be integers, not 'str' on line 31" from time to time, possibly because of network error so the models list couldn't load and method was trying to parse this data. Now the code is updated in order to check for any potential errors first.

* Update BlackboxAPI.py: remove unused imports 

format_prompt() and JSON library are not being used here, so they may be removed safely.

* Update copilot.yml

This job is failing due to the error in JavaScript code; this commit is trying to fix it.

* Update providers-and-models.md to reflect latest changes

Updated models list and removed providers that are currently not working.